Refactor FS RNG files into rng folder
