+++ marlin/src/parameters.rs
@@ -14,12 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{
-    marlin::{FiatShamirChaChaRng, MarlinSNARK},
-    ProvingKey,
-    VerifyingKey,
-    SRS,
-};
+use crate::{marlin::MarlinSNARK, rng::FiatShamirChaChaRng, ProvingKey, VerifyingKey, SRS};
 use snarkvm_algorithms::errors::SNARKError;
 use snarkvm_curves::traits::{AffineCurve, PairingEngine};
 use snarkvm_r1cs::ConstraintSynthesizer;
