+++ marlin/src/rng/traits/algebraic_sponge.rs
